refactoring NavBar out into main App component
